third_party/WebKit/LayoutTests/media/track/track-webvtt-tc007-cue-no-id.html

<!DOCTYPE html>
<title>Tests empty cue identifiers (they are optional), but makes sure "-->" found leads to discarded cue.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
check_cues_from_track("captions-webvtt/tc007-cue-no-id.vtt", testTrack);
check_cues_from_track("captions-webvtt/tc007-cue-no-id-error.vtt", testTrack);

function testTrack(track) {
    var expected = [
        { id: "", startTime: 0, endTime: 30.5, text: "Bear is Coming!!!!!" },
        { id: "", startTime: 31, endTime: 60.5, text: "I said Bear is coming!!!!" },
        { id: "", startTime: 61, endTime: 1200.5, text: "I said Bear is coming now!!!!" }
    ];

    assert_cues_match(track.cues, expected);
}
</script>
